Refactor model loading and introduce Vivit model wrapper with optional projection head
